Verify Project Setup with Development Command
This pull request aims to verify the project setup by running the development command. The project uses a combination of TypeScript, React, and Tailwind. To ensure the project is set up correctly, we need to run the command specified in the `package.json` file under the `scripts` section. If the `dev` script is specified, we will use the corresponding command (`npm run dev` or `yarn dev`) to start the development server.